Ground Rotation and Slide Detection Fixes
Ground Rotation has a raycast in the back, this is for the case when the
front raycast does not detect ground. EX: When jumping onto a slope
facing downwards, you will now rotate downwards

Slide Detection back raycast has been adjusted, it goes with the slide
so that regardless of the squirrel's rotation, it will detect when you
shouldn't be sliding anymore.

Removed Slide Rotation Constraint